poplar: add support for USB recovery
When the Poplar board boots it records values in some registers that
can be used to determine whether the board was booted with the
USB_BOOT button pressed.  We will use that to determine whether
we're doing a boot truly intended to be a USB recovery operation
(i.e. reinitialization/de-brick).

We add a last-stage initialization function that examines these
registers to see if we're doing a recovery boot.  If not, we skip
this step and proceed with a normal boot.

If it is a recovery boot, display a prominent message, and then ask
the user if they really want to proceed with reinitializing things.
Unless they say "yes", we abort the recovery and proceed with normal
boot.

Otherwise, we initiate the recovery process.  A command is issued to
to load the installer script from the USB media and run it (using
the "source" command).  This script formats the eMMC with a
partitioned image for running Linux.

If a recovery operation begins, it will complete with a reset
(whether it succeeded or not).

Signed-off-by: Alex Elder <elder@linaro.org>